Cleanup, Python3-compatible raise statements
